allow.all: Do not regard libvirt guest start failures as regressions
We have been experiencing some regressions in test-amd64-*-libvirt
guest-start.

These appear to be host-specific: for example, test-amd64-i386-libvirt
succeeded in 34278 on itch-mite but failed in 34234 on bush-cricket.
The only difference in osstest's records of the versions is a change
to linuxfirmware.

I booked out field-cricket and have repro'd some problems there.  It
appears that libvirtd can crash.  The symptoms are not the same as in
the push gate failures in osstest (34316, 34234) and linux-linus
(34299) but similar to those in linux-3.0 (34291) and linux-3.16
(34285).

When running libvirtd in the foreground we can get this:
  libvirtd: libxl_event.c:1791: libxl__ao_complete_check_progress_reports: Assertion `ao->in_initiator' failed.

So, it appears to be some kind of race.  The libvirt tests are new so
this is not a real regression in any of the trees gated by osstest.
So, permit this test for now.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>