allow files: Support negation syntax
We now support !<pattern>, which undoes the effect of a <pattern>
specified previously (perhaps in another file).

Note that the logic is a bit poor: !... only undoes the effect of an
identical (or equivalent) earlier pattern; it doesn't provide a
specific override.

Improving that would involve a more invasive change to
sg-report-flight which I don't want to do right now.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>