PDU: pdu-msw: Split $read_oid and $write_oid
Some PDUs have a different OID space for turning ports on and off, to
the one for querying.  To make this easier to handle, split the
variable $oid into $read_oid and $write_oid.

Also move $baseoid settings earlier so that we can modify them with
command-line arguments.

No functional change in this patch.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>