PDU: pdu-msw: Support APC v6 firmware
APC PDUs with firmware 6.x have a different OID space for turning
ports on and off, to the one for querying.  The old namespace still
works to turn the port on and off but returns a genErr error response!

Support a new command-line option --apc6 to use this other OID.

Information about OIDs etc. obtained here:
  https://git.fedorahosted.org/cgit/fence-agents.git/commit/?id=55ccdd79f530092af06eea5b4ce6a24bd82c0875

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>