allow: Allow rumpuserxen-demo-xenstorels/xenstorels failures on all branches
This test is subject to a heisenbug. The next step after this one is a
repetition (50 times) of the same test which should reliably hit the
issue.

In the meantime this individual attempt sometimes hits and sometimes
doesn't, causing unhelpful false regressions.

The repeated test has never passed in any branch and so is not yet
considered a regression.

Since the repeated test will catch the actual heisenbug I haven't
bothered to add a rumpuserxen specific override which makes this step
blocking again.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>