cr-ensure-disk-space: Permit argument to specify local directory.
If the argument is Logs rather than LogsPublish (ie, refers to a local
directory (without `:') rather than a remote one (with `:'), do things
locally (by invoking sh -ec so that we have identical quoting rules to
ssh).

No effectively functional change with the current configuration.

We still always do a local deletion.  This is anomalous and will
disappear shortly.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>