uboot: do not use readlink on xsm policy
The policy is not a symlink, so readlink will return nothing.

We cannot use readlink -f because that will return an absolute path
and we need a path relative to the filesystem root (in this case
/boot).

Keep flaskpolicy=$flaskpolicy as a shell variable rather than
unescaping the uses (so they are interpreted by Perl) to easy any
future changes.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>