target_cmd_build: Delete build-ok-stamp before starting
Many of the callers of target_cmd_build use a build-ok-stamp idiom to
detect failed builds.  This idiom does not work if the stamp file
exists already, so delete it.

In the future we may move more of the test build-ok-stamp, echo ok,
into TestSupport, but this will do for now.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>