cambridge: Switch configuration to use osstest.xs.citrite.net
The VM has moved to different infrastructure and its new name is
osstest.citrite.net.

Update ExecutiveDbnamePat. The DB is still in the XC infrastructure so
using DnsDomain (the default) no longer works.

Set {Owner,Queue}DaemonHost to refer to the new VM host and not the
default ControlDaemonHost value of control-daemons.osstest.cam.xci-test.com
(which will be removed later).

We set both variables rather than just ControlDaemonHost in case we
ever want to move one but not the other.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>