cs-bisection-step: Use pbm tools, not graphicsmagick/imagemagick
Graphicsmagick / imagemagick have very poor performance with images
with large pixel sizes.  The bisector can generate some very large
images.

In an example I have seen, a 21595x21048 png, occupying only 2.6Mby of
disk space.  An invocation of `convert' to resize this was using 3Gby
of RAM and lots of CPU.  Whereas, the pbm utilities can process this
with much less memory and a tiny fraction of the cpu time.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>