sg-report-flight: Report stepno and testid of first worst fail
This makes reading the scoreboard considerably easier.

We abuse the local variable @worst slightly, pushing the extra info we
are going to print onto the end of it.

We also have to defer printing the cells, because we compute the cell
to duplicate in column order but we have to output them in row order.
For symmetry we accumulate both rows rather than only the second row.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>