Debian: refactor code to add preseed commands to the preseed file
Call it from ts-debian-hvm-install.

This means that, in future, ts-debian-di-install can use
preseed_hook_command and preseed_hook_installscript.

The existing opencoded use of d-i/late_command in the guest preseed
needs to become a preseed_hook_command so as not to clash with the use
of preseed_hook_cmds().

This requires also adding a #! line and the "set -ex" boilerplate
which in turn requires slightly rewriting the /boot/EFI handling part
to also work if the system is not installed for EFI (in which case
grubx64.efi isn't installed). Previously this would have needlessly
created the directory and then ignored the error from cp.

The ssh authorized keys bit isn't touched since it works as is and
will go away in a subsequent patch.

Apart from no longer creating /target/boot/efi/EFI/boot when it is not
needed there is no functional change for now.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Cc: longtaox.pang@intel.com
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>