ts-debian-hvm-install: Only apply EFI workaround for Wheezy
The previous refactoring of preseed hooks makes this easy to do.

The underlying issue is lack of persistent variable store in our OVMF
setup, which we workaround by placing a copy of grub at the removable
media path. Add a comment saying this since I initially thought this
was just a Wheezy bug.

In Jessie the extra copy of grub can be achieved by preseeding
grub-installer/force-efi-extra-removable (since various real h/w has
similar limitations/bugs) however I haven't tested that so I didn't
add it to the preseed yet, I just mention it in the code comment.

Currently this script hardcodes Wheezy, refactor to use the 'suite'
guest_var (or $c{GuestDebianSuite})

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Cc: wei.liu2@citrix.com
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>