cambridge: Reduce reliance on woking and configdb
The hosts which were previsouly attached to woking's serial ports are
now attached to osstser1 which is our own box.

All hosts which previously used configdb (and by extension statedb)
are now configured directly via the host properties entries for PDU
control (direct to PDU) and for Ethernet address, so disable configdb.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>