reporting: sg-report-flight should ignore missing jobs
The function findaflight should not, when the caller specifies a job,
find a flight which does not contain that job at all.

One effect of allowing it to find such flights is that it might find a
bisection flight and try to use it as a basis for claiming a
regression, or as a justification for something not being a
regression, and then complain that all the missing steps in the
bisection flight are `blocked'.

This can be seen in the report for 58627:
 test-amd64-i386-xl-qemuu-winxpsp3  6 xen-boot fail blocked in 56366-bisect

After this patch, a report generated for 58627 no longer mentions
56366 at all.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>